requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
distro==1.9.0
exceptiongroup==1.2.1
flatbuffers==24.3.25
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
joblib==1.4.2
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.0
openai==1.32.0
openwakeword==0.6.0
packaging==24.0
protobuf==5.27.1
PyAudio==0.2.14
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
sniffio==1.3.1
soupsieve==2.5
sympy==1.12.1
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
